updating class performance report API to serve the data weekly and montly with filter on subject and framework

+++ src/main/java/org/gooru/groups/reports/country/perf/GroupReportByCountryResponseModel.java
@@ -18,20 +18,20 @@ public class GroupReportByCountryResponseModel {
   }
 
   static class StateLevelPerf {
-    private Integer id;
+    private Long id;
     private String name;
     private String code;
     private Long timespent;
     private Double performance;
 
-    public Integer getId() {
+    public Long getId() {
       return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
       this.id = id;
     }
-
+    
     public String getName() {
       return name;
     }
